TG: index.js content split up into index.js and app.js files
