+++ src/app.js
@@ -1,0 +1,14 @@
+import express from 'express';
+import router from './routes';
+
+const app = express();
+
+app.use(express.json());
+
+app.get('/', async (req, res) => {
+  res.send('Hello World!');
+});
+
+app.use(router);
+
+export default app;
