+++ src/index.js
@@ -1,16 +1,11 @@
-import dotenv from 'dotenv';
-import express from 'express';
 import mongoose from 'mongoose';
-import router from './routes';
+import dotenv from 'dotenv';
+import app from './app';
 
 dotenv.config();
 
-const app = express();
 const port = process.env.PORT || 3000;
 const URL = process.env.DATABASE_URL;
-
-app.use(express.json());
-app.use(router);
 
 mongoose.connect(URL, (error) => {
   if (error) {
@@ -19,14 +14,9 @@ mongoose.connect(URL, (error) => {
   } else {
     // eslint-disable-next-line no-console
     console.log(':círculo_verde_grande: Database connected');
+    app.listen(port, () => {
+      // eslint-disable-next-line no-console
+      console.log(`Example app listening on port ${port}`);
+    });
   }
-});
-
-app.get('/', async (req, res) => {
-  res.send('Hello World!');
-});
-
-app.listen(port, () => {
-  // eslint-disable-next-line no-console
-  console.log(`Example app listening on port ${port}`);
 });
